refactor NetworkDiagnosticsButton, split out sub sections
